[OS][ww44] - Upstreaming compiler repository (#40)
* [OS][ww44] - Upstreaming compiler repository

* [OS] Update submodules

* [OS] Update schema paths

* [OS] Upstreaming compiler repository - remove upa

+++ src/vpux_utils/src/core/unix/memory_usage.cpp
@@ -7,6 +7,7 @@
 
 #include <fstream>
 #include <regex>
+#include <sstream>
 
 namespace vpux {
 
@@ -23,7 +24,7 @@ vpux::KB getPeakMemoryUsage() {
             iss >> peakMemUsageKB;
         }
     }
-    return vpux::KB(peakMemUsageKB);
+    return vpux::KB(static_cast<int64_t>(peakMemUsageKB));
 }
 
 }  // namespace vpux
